refactor(field): extract field-type static seam off the dead item layer (C-24 train 1, behaviour-preserving)
FieldTypeManager resolves every field-type plugin by calling static methods on its
class (defaultSettings/defaultValue/schema/jsonSchema/jsonSchemaFor/schemaFor) — it
never instantiates a plugin. Those statics lived on FieldItemBase, which also
carries the DEAD Field-API item/value-object layer (ComplexData/TypedData instance
methods, validate() no-op, PropertyValue) instantiated nowhere in production.

Train 1 moves the static seam into a new @internal abstract base AbstractFieldType;
FieldItemBase now extends AbstractFieldType, so the seam is single-sourced and
INHERITED UNCHANGED. No field-type plugin is reparented and no public symbol is
removed yet — that is train 2 (reparent + delete the instance layer + its tests);
train 3 retires the now-unused typed-data ancestry. Pre-1.0-alpha: trains 2-3 break
the dead @api symbols rather than shimming.

Behaviour-preserving: new FieldTypeStaticSeamResolutionTest asserts every discovered
field type still resolves through the seam and reaches it via AbstractFieldType; the
full field suite (480 tests) stays green with zero test changes. surface-parity OK
(AbstractFieldType mapped internal); cs/phpstan/symfony-imports/layers green.

Plan: docs/audits/c24-field-item-layer-deprecation-scoping.md.
