+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- **The live field-type "descriptor" static seam was extracted off the dead Field-API item layer (audit C-24, train 1 of 3 — behaviour-preserving).** `FieldTypeManager` resolves every field-type plugin by calling static methods on its class (`defaultSettings`/`defaultValue`/`schema`/`jsonSchema`/`jsonSchemaFor`/`schemaFor`) — it never instantiates a plugin. Those statics lived on `FieldItemBase`, which also carries the **dead** Field-API item/value-object layer (the `ComplexData`/`TypedData` instance methods, the `validate()` no-op, `PropertyValue`) that is instantiated nowhere in production. Train 1 moves the static seam into a new `@internal` abstract base `Waaseyaa\Field\AbstractFieldType`; `FieldItemBase` now `extends AbstractFieldType`, so the seam is single-sourced and **inherited unchanged** — no field-type plugin is reparented and no public symbol is removed yet. Acceptance: new `FieldTypeStaticSeamResolutionTest` (every discovered field type still resolves through the seam and reaches it via `AbstractFieldType`) + the full field suite (480 tests) stays green. **Train 2** reparents the field-type plugins directly onto `AbstractFieldType` and removes the dead instance layer (`FieldItemList`/`FieldItemBase` instance methods/`PropertyValue`) + the tests that exercise it; **train 3** retires the now-unused `typed-data` ancestry. Pre-1.0-alpha: trains 2–3 break the dead published `@api` symbols rather than shimming. Plan: `docs/audits/c24-field-item-layer-deprecation-scoping.md`.
+
 ### Added
 
 - **Authorized download path for private attachment bytes — `GET /attachment/{id}/download` (CL-13, slice 1 of 2).** Entity-row access policies protect the attachment *record*; previously nothing in the framework served its *bytes*, and the `public://` + `/files/` convention encouraged hosts to serve `storage/files` statically (which would expose any attachment's bytes by URL regardless of the entity policy). This adds the safe serving path: a new `private://` scheme on a non-served root (`storage/private-files`, deliberately a sibling of `storage/files` so a `/files` static mount can't reach it — override via `config['attachment']['private_files_root']`), resolved through `Waaseyaa\Attachment\Storage\PrivateFileStore` with `realpath()` containment (defeats `../` traversal + symlink escapes). `AttachmentDownloadRouter` (a foundation `DomainRouterInterface`, matched via the `attachment.download` controller string registered in `BuiltinRouteRegistrar` next to `media.upload` — an L2 package can't depend on routing/L4) is the enforcement point: it runs the **deny-by-default** `view` check — which `ParentDelegatedAccessPolicy` delegates to the owning parent entity — and only then streams the bytes; an unauthorized caller, a missing/unbound account, a non-`private://` `storage_uri`, or a traversal URI all return **404** (never 403 — no existence oracle). Public assets keep `public://` and the host's normal static path, unaffected. Acceptance: `AttachmentDownloadRouterTest` (authorized caller streams the bytes; unauthorized/no-account/non-private/traversal all 404 with no bytes). **Slice 2 (media, follow-up)** adds the same authorized download for `media`, resolving the file URI through the media type's source plugin (the single source of truth — not a duplicate file-uri field).
