+++ docs/specs/entity-system.md
@@ -1,5 +1,6 @@
 # Entity System
 
+<!-- Spec reviewed 2026-06-23 - audit C-24 train 1 (field item-layer deprecation): behaviour-preserving internal refactor of the field package, no contract/behaviour change. The live static field-type "descriptor" seam (defaultSettings/defaultValue/jsonSchemaFor/schemaFor) was extracted off the dead FieldItemBase instance/value-object layer into a new @internal abstract base Waaseyaa\Field\AbstractFieldType (which FieldItemBase now extends, so the seam is single-sourced and inherited unchanged). FieldTypeManager still resolves every field-type plugin via $class::method() — identical output, proven by FieldTypeStaticSeamResolutionTest (all discovered types resolve + reach the seam via AbstractFieldType) and the full field suite (480 tests). No field-type plugin reparented yet and no surface removed — train 2 reparents the plugins onto AbstractFieldType and removes the dead instance layer + its tests; train 3 retires the now-unused typed-data ancestry. Scoping: docs/audits/c24-field-item-layer-deprecation-scoping.md. -->
 <!-- Spec reviewed 2026-06-23 - audit C-6 (#1714 follow-through): the entity query layer is now deny-by-default. SqlEntityQuery::filterCandidates() survivor test flipped from !isForbidden() (admitted Allowed AND Neutral) to isAllowed(), so a Neutral row ("no policy granted a view") is dropped, not admitted; count() inherits this. The empty/unwired-handler fallback now denies every candidate (fail-closed) instead of returning the unfiltered window — closing the open-by-default primitive. Production wiring (the #1714 accessHandlerResolver) is unchanged; the fallback is reached only by direct construction or a missing resolver. The full enforcement-layer contract lives in docs/specs/access-control.md (Layer 3, updated in the same PR); entity-system pipeline/storage/save-load contracts are unchanged. Genealogy SSR (a surfaced pre-existing prod defect) now gathers relationship topology via accessCheck(false) with per-person gate concealment so living/private relatives render as redacted placeholders rather than vanishing. Acceptance: SqlEntityQueryDeniesNeutralRowTest + GraphQL/discovery/semantic/genealogy integration suites. -->
 <!-- Spec reviewed 2026-06-22 - WP16 (alpha245 security, issue #1714): SqlEntityStorage::getQuery() is now fail-closed in production wiring. resolveAccessHandler() falls back to an EMPTY EntityAccessHandler when none is wired (check() returns Neutral → !isForbidden() admits every row), and both build sites (AbstractKernel storage closure, EntityStorageFactory) built storage with accessHandler=null → fail-OPEN. SqlEntityStorage gains an optional `?\Closure(): ?EntityAccessHandler $accessHandlerResolver`; getQuery() resolves the handler from `accessHandler ?? resolver()` at QUERY time (an explicit handler still wins) — resolving lazily means a storage built/cached mid-boot, before the kernel's handler is populated, still wires the real handler at runtime. EntityStorageFactory gains the same optional resolver param (default null = unfiltered, system-context only). accessCheck(false) bypass and the missing-account fail-closed throw are unchanged. Acceptance: EntityStorageFactoryAccessWiringTest::{factory_wired_storage_getQuery_filters_forbidden_rows, unwired_factory_storage_is_fail_open_documenting_the_pre_fix_gap} + the kernel-booting integration suite (962, WP05/WP06 access paths unaffected). -->
 <!-- Spec reviewed 2026-06-22 - WP13 (classification security, audit #18/#19): two field/classification fixes. (1) PurgeJob now enforces a hard legal-hold guard — a `hold-*` classification label is NEVER purged, regardless of policy match. The class doc asserted this held "by construction" (purge policies use action=purge, holds use action=hold-flag), but a misconfigured purge policy whose `applies_to` glob matched `hold-*` (e.g. `hold-*` or `*`) would otherwise delete legal-hold content; the invariant is now enforced in code (`str_starts_with($labelId, 'hold-')` skip) so it cannot be configured away. (2) FieldServiceProvider now wires the `classification.role_clearance` config into RoleBasedClearanceChecker — the binding previously did `new RoleBasedClearanceChecker()` with no argument, so the host override was silently ignored and clearance was permanently the stock DEFAULT_ROLE_CLEARANCE; the checker still falls back to that default when no override is present. Acceptance: PurgeJobTest::never_purges_legal_hold_labelled_entities_even_when_a_policy_matches + FieldServiceProviderClearanceConfigTest::honours_the_classification_role_clearance_override (both proven to fail against the pre-fix code). -->
